reorg test and fix linter warning
