+++ tests/unit/tabular/test_curation_status.py
